refactor: replace Ops with a proper AST
Introduce an AST that closely maps to the QBE IR to simplify code
generation. This is still quite cumbersome and supports only two
instructions (`call` and ret`). I'll have to see how this develops.

Note: I don't think we keep enough information for proper type checking,
so this will need to be enhanced as well (e.g. while we parse the
`extern` keyword, we throw away the type definition and don't check the
function calls)